+++ src/lib.rs
@@ -2,5 +2,6 @@ pub mod args;
 pub mod clap;
 pub mod commands;
 pub mod core;
+pub mod draw;
 pub mod errors;
 pub mod style;